backend/pyproject.toml
[tool.poetry]
name = "help-me-doctor-backend"
version = "0.1.0"
description = "HelpMeDoctor — Singapore Medical/Legal RAG Multi-Agent Backend"
authors = ["HelpMeDoctor Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = { extras = ["standard"], version = "^0.30" }
langgraph = "^0.2"
langchain-google-vertexai = "^2.0"
langchain-core = "^0.3"
motor = "^3.5"
beanie = "^1.26"
pydantic-settings = "^2.4"
python-jose = { extras = ["cryptography"], version = "^3.3" }
passlib = { extras = ["bcrypt"], version = "^1.7" }
python-multipart = "^0.0.9"
pandas = "^2.2"
google-cloud-aiplatform = "^1.60"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8"
pytest-asyncio = "^0.23"
httpx = "^0.27"
ruff = "^0.5"
mypy = "^1.11"
motor-types = "^1.0"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
